requirements.txt
pynput==1.7.6
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
